backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.0
Flask-Migrate==4.0.5
cryptography==41.0.7